[chore] authTypes -> Type.ts 파일명 변경
