+++ Client/src/components/form/FindPasswordForm.tsx
@@ -3,7 +3,7 @@ import Button from '../ui/Button'
 import Input from '../ui/Input'
 import ErrorText from '../ui/ErrorText'
 
-import { AuthFormData } from '../../types/authTypes'
+import { AuthFormData } from '../../types/Types'
 
 import { validateEmail, validatePassword, validateConfirmPassword } from '../../utils/validation'
