+++ Client/src/components/form/LoginForm.tsx
@@ -3,7 +3,7 @@ import Button from '../ui/Button'
 import Input from '../ui/Input'
 import ErrorText from '../ui/ErrorText'
 
-import { AuthFormData } from '../../types/authTypes'
+import { AuthFormData } from '../../types/Types'
 
 import { validateEmail } from '../../utils/validation'
 import { Link, useNavigate } from 'react-router-dom'
