+++ Client/src/pages/Login.tsx
@@ -1,7 +1,7 @@
 import { useNavigate } from 'react-router-dom'
 import Block from '../components/ui/Block'
 import { useState } from 'react'
-import { AuthFormData } from '../types/authTypes'
+import { AuthFormData } from '../types/Types'
 import useAccountStore from '../stores/AccountStore'
 import { requestLoginUser } from '../api/userApi'
 import LoginForm from '../components/form/LoginForm'
