+++ Client/src/types/Types.ts
@@ -1,0 +1,15 @@
+export interface AuthFormData {
+  email: string
+  password: string
+  confirmPassword: string
+  apiKey: string
+  character: string
+}
+
+export interface ExpeditionCharacter {
+  ServerName: string
+  CharacterName: string
+  CharacterLevel: number
+  CharacterClassName: string
+  ItemAvgLevel: string
+}
