--- Client/src/types/authTypes.ts
@@ -1,7 +1,0 @@
-export interface AuthFormData {
-  email: string
-  password: string
-  confirmPassword: string
-  apiKey: string
-  character: string
-}
